Able to complete player class Added deck(int array) which goes from 1-10 to reg numbers. Along with variables with speciality cards(king,queen,jack, and aces)
first two cards dealt to player are randomized and are added

userhit function:
displays the first two card
displays the total of the first two cards

proceeds to a while loop:
randomnumber variable displays the card that has just been added to the deck

Asking if user hits, the loop checks to see if it is a specialty card and announces the speciality card(king,queen,jack). If  it's an ace, user has choice to do either a 1 or 11

In general, the value is added to the total.

If user stands, the total is displayed.

if the user total equals 21, it is a blackjack, if it's greater than 21, bust.

TODO:
start programming computer class
Optimize/debug/polish player class